requirements.txt
boto3==1.17.22
botocore==1.20.22
dj-static==0.0.6
Django==2.2.19
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
djongo==1.3.4
jmespath==0.10.0
pymongo==3.11.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.2.4
static3==0.7.0
urllib3==1.26.3
